Step 4: Name dotfiles install script 'install.sh'
Apparently, and as opposed to [what is documented][dotfiles], a file
named 'setup.sh' does not work to direct the installation of dotfiles
into a newly created Codespace.  While the GitHub Codespaces team works
to fix that bug, let's have students instead name the file 'install.sh',
which does seem to work.

[dotfiles]: https://docs.github.com/en/codespaces/setting-your-user-preferences/personalizing-github-codespaces-for-your-account#dotfiles